Defer transfers until draw
Before we did a transfer as soon as it was requested which has a lot of
overhead. Now we wait until frame draw before performing all transfers
at once. This seems to introduce some race conditions or make existing
one worse, but this should be fixed instead of discarding this nice
change :)